修改网络名称以及参数，并且适配可以预加载paddle的预训练模型
